requirements.txt
pytest==9.0.1
pytest-mock==3.15.1
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.44
sentence-transformers==5.1.2
